Implement jet PODIO relations and omnify factories (#1067)
### Briefly, what does this PR introduce?

This PR brings the jet reconstruction algorithm and factories up-to-date
with the reconstruction framework by (1) linking jets to their
constituents via PODIO relations (in contrast to the
[previous](https://wiki.bnl.gov/EPIC/index.php?title=JetsHF) approach),
and (2) converting existing factories to `JOmniFactory`.

In doing so, the jet factories have been vastly simplified to the point
where both the Generated and Reconstructed Jets can make use of the same
factory.

- [x] Requires PR #1217 first.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #902 )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [X ] Documentation has been added / updated (the Jet/HF wiki page will
be updated on merging)
- [X ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

Yes. Any code making use of the generated and reconstructed jet
collections will need to be updated to

1. avoid checking the jet `type` for if the object is a jet or
constituent; and
2. instead refer to constituents via PODIO relations.

### Does this PR change default behavior?

Yes. Now, only jets will be stored in the output collections, and the
`Generated*Jets` collections are constructed by processing the
`GeneratedParticles` collection.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>