feat: register DD4hep_service as algorithms::GeoSvc provider (#1195)
To be able to use the algorithms interface, including the service
requirements defined upstream, we need to connect the EICrecon
DD4hep_service to algorithms::GeoSvc as a provider. This does exactly
that.

The new AlgorithmsInit_service depends on the DD4hep_service,
Log_service, and we use the algorithms RandomSvc since we don't have one
in EICrecon yet.

TODO:
- [x] ~~depends on https://github.com/eic/algorithms/pull/14~~ (or not
really, but would avoid a const_cast)
- [x] depends on https://github.com/eic/algorithms/pull/16

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>