feat: add cmake-format and cmake-lint support and first pass cleanup (#1147)
### Briefly, what does this PR introduce?
This cleans up the formatting of all cmake files, and adds a linter
through pre-commit. It follows the standard conventions of cmake-format
with minimal exceptions: line length is set to 120 for now (didn't want
to chop at the default 80 since it changes many lines in ways that are
more than whitespace), and function names are lowercase (per
https://cliutils.gitlab.io/modern-cmake/chapters/intro/dodonot.html
recommendations of keeping uppercase for global variables). A few
pragmas are allowing exceptions, mainly about capitalization of certain
variables).

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1141)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.