Associate matched calo cluster with reco electrons (#1276)
### Briefly, what does this PR introduce?
The previously introduced ElectronReconstruction factory outputs a
ReconstructedParticleCollection of electron candidates via the E/p
ratio. However, the matched cluster (used for the E/p calculation) was
not associated. This calls repo_electron.addToClusters(...) to keep the
association for use in later electron finder factories that use the
electron candidates.

Note:
I also considered adding the energy (from the cluster) to the reco
Electron via:
```
reco_electron.setEnergy(cluster.getEnergy());
```
but I am not sure if this is a good idea...

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__) 
- [ ] New feature (issue #__)
- [ ] Documentation update
- [x] Other: __ Something between a bug fix and feature update

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
NO

### Does this PR change default behavior?
NO

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>