Fix Jet Algorithm Initialization Bug (#1331)
### Briefly, what does this PR introduce?

This PR resolves issue #1254. The `JetReconstruction` algorithm now
determines the number of parameters needed for the selected fastjet
algorithm and calls the relevant `JetDefinition` ctor.

In the process of resolving the issue, a new user-configurable parameter
has been introduced into the algorithm's configuration: `pJet`, which
sets the exponent on the energy weights of the distance measure in the
case of the generalized kt algorithms.

### What kind of change does this PR introduce?
- [X] Bug fix (issue #1254 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No.

### Does this PR change default behavior?

No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>