treewide: use PODIO ObjectID instead SimID/RecID indices (#1346)
### Briefly, what does this PR introduce?

This is a subset of #970, that removes use of SimID and RecID fields
(while switching to comparisons of full object id's instead of just the
indices).
In principle, we could even drop `.getObjectID()` in comparisons and
just use `Object::operator==`
https://github.com/AIDASoft/podio/blob/400f4f06bef3e7666c71e31be579ec715c412b8a/python/templates/macros/implementations.jinja2#L160,
which relies on pointer comparison instead.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
No