fix: ensure eicrecon-this.sh prepends even when already in PATH (#1205)
### Briefly, what does this PR introduce?
`eicrecon-this.sh` had an issue where it would **not** add EICrecon
installs to the PATH and LD_LIBRARY_PATH if they were already in there,
**even if there was another EICrecon install earlier in those paths**.
That results in confusion when juggling (no pun intended) multiple
worktrees with various EICrecon installs.

This PR adds a small utility `pathadd` that takes care of three things:
- always adds the specified directory to the front of the environment
variable list to make sure it is found first,
- checks if the environment variable you want to add to is already there
or not to avoid trailing colons,
- removes the specified directory from the interiors of the path if it
is already somewhere.

On top of that it's POSIX-safe and doesn't use bashisms (which is what
took 50% of the effort; writing a regex match and replace is the easy
way).

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No. (Default behavior is having just one install tree...)

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>