Adapt ElectronReconstruction to use MatchClusters (#1355)
### Briefly, what does this PR introduce?
The ReconstructedElectron algorithm now uses the particle-cluster
associations created by MatchClusters_factory rather than doing it
itself. It makes the collection ReconstructedElectrons a subset
collection of ReconstructedParticles, so the association is kept between
them.

I will leave this as a draft until #1277 is merged, then update that
code as well to take advantage of it.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #1320 )
- [ ] New feature (issue #__ )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
ReconstructedElectrons is now a subset collection, but I don't believe
that should break anything.

### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>