Remove m_ member tags from tracking configuration structs (#1377)
### Briefly, what does this PR introduce?
Tidies the code so that the tracker configuration structs don't have the
m_ tag. None of the other algorithm configuration structs are prefixed
by m_ and it is my understanding that these are conventionally for class
members which presumably they may have originally been before moving to
a separate config.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [x] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
No