fix(ci): add src/tests plugins as CI jobs (#1386)
### Briefly, what does this PR introduce?
We don't test what's in src/tests, to a large degree (exception is the
catch unit tests), and this leads to code issues, e.g. #1385.

This PR adds the test plugins in src/tests to a CI job. ~~It won't
actually catch #1385 until we stop silently eating exceptions...~~ It
catches #1385 now,
https://github.com/eic/EICrecon/actions/runs/8726213766/job/23941043650#step:8:676.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: failures like #1385 are not caught)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.