LUMISPECCAL: LumiSpecCALHits -> EcalLumiSpecHits (#1381)
### Briefly, what does this PR introduce?

The LumiSpecCAL name is from deprecated
`$DETECTOR_PATH/compact/far_backward/lumi/spec_homo_cal.xml` that was
replaced by the `spec_ScFi_cal.xml` in
https://github.com/eic/epic/commit/eaf7f8089c16ef4e935e97c671e957f82942fea4#diff-9ff1b47fbddbe9afe982f6dc6cce5c5f43a15362ff43fa4216c0b4f34a4bd72eR10.

This PR brings EICrecon in accordance to the geometry change. The sector
field is deprecated and is removed as part of this change.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
Yes