Add filter by association templated factory (#1326)
### Briefly, what does this PR introduce?
Adds a factory which creates a subset collection containing only members
which are associated to the contents of another collection.

The motivation and example is in finding the reconstructed electron
which is associated with the scattered electron. This is currently
conducted independently in several of the reconstruction algorithms
using copied code. Having a common collection will stop this being run
several times. An alternative would be providing a single function which
would clean up the code but still require running multiple times for the
same result.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>