fix(ci): run at least compilation and unit tests on few fixed jug_xl (#1136)
### Briefly, what does this PR introduce?
This changes the CI philosophy to run the compilation and unit testing
on an expanded set of platform versions. It does this for the clang
release versions, since fastest. It separates the default CI platform
(set as input on workflow_dispatch) from the explicit platform versions
sets in the build job matrix (there may be overlap, even by default).

### What kind of change does this PR introduce?
- [x] Bug fix (issue: need to test on more versions/platforms; have
tests on stable platform by default)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>