fix: pass logger options from factory to algorithms-interfaced algorithms (#1407)
### Briefly, what does this PR introduce?
With the switch to the algorithms interface, we need to explicitly pass
the logger level from the JOmniFactory to the algorithm, i.e.
`m_algo->level(static_cast<algorithms::LogLevel>(logger()->level()));`.
Only this allows the algorithm to use the log level that is specified on
the command line.

This PR adds the necessary line for those factories that use
algorithms-interfaced algorithms. There may be other factories that I
overlooked. Not sure how to fail compilation when we don't do this.

### What kind of change does this PR introduce?
- [x] Bug fix (issue: `-PDRICH:DRICHRawHits:LogLevel=trace` ignored)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.