feat: convert IrtCherenkovParticleID factory to JOmniFactory (#1418)
### Briefly, what does this PR introduce?
This PR converts the IrtCherenkovParticleID to JOmnifactory, and the
IrtCherenkovParticleID algorithm to the algorithms interface.

Because the IrtCherenkovParticleID factory was doing some parsing of
collection names, there is less flexibility for now. In particular, for
now, the IrtCherenkovParticleID factory temporarily hardcodes the DRICH
as the RichGeo system.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.