feat: set track relation for calorimeter and DRICH propagations (#1450)
### Briefly, what does this PR introduce?
The tracks propagated to the calorimeter surfaces and the DRICH aerogel
and gas surfaces did not contain the relation to the track.

This PR adds the track relations to CalorimeterTrackProjections and the
DRICH{Aerogel,Gas}Tracks collections.

Note the methodology here. We don't use edm4eic::Tracks to do the
propagation, so this here relies on keeping the order in the
edm4eic::Tracks and ActsExamples::Trajectories.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue: no track relations in projections)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.