fix: don't double-define logger LogLevel parameters (#1436)
### Briefly, what does this PR introduce?
The Log_service::logger(name) call defines name:LogLevel already, so
callers don't need to do this anymore. That removes the need for
SpdlogExtensions.h includes in factories.

This PR removes the double defines, removes the headers, and also
unifies the Acts loggers, so we don't have one for init and one for
regular.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.