Cluster axis direction (#1391)
Added code to produce the direction of clusters in
CalorimeterClusterRecoCoG

### Briefly, what does this PR introduce?
adds the cluster direction (i.e., the eigenvector corresponding to the
largest eigenvector of the matrix corresponding to the inertia matrix of
the cluster)

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X] New feature (issue #1462 )
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
no
### Does this PR change default behavior?
Yes. While calculating the shape parameters, the hits are assigned
weights, in a similar manner to how they are assigned weights when
determining the cluster position. Previously, there were some
inconsistencies in how these weights were determined when using certain
options in the cluster fitting, which were fixed in order to make this
feature (the cluster axis direction) work properly.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>