fix: convert to algorithms logger, no m_log in each algorithm class (#1483)
### Briefly, what does this PR introduce?
This PR moves a lot of algorithms and factories over to using the
internal logger functionality that comes with the algorithms interface.
It means that individual algorithms don't need to initialize their own
`m_log`. It also means that these algorithms are now properly conformant
to the algorithms interface, i.e. `init()` and `process(input, output)
const`.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.