Inject Low-Q2 electrons into tracking (#1491)
### Briefly, what does this PR introduce?
Injects the LowQ2 tracks into the tracking workflow so they can finally
be used like any other particle.

This might not be the best place to add the information from the events
in the longer term as there will may be some extra looping over
calorimeter and pid hits which aren't possible. At the moment though
there isn't another easy place to add them while there is a mix of
ReconstructedChargedParticles and ReconstructedParticles used in
different places.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No

### Does this PR change default behavior?
Low-Q2 electrons should show up in ReconstructedParticle collections and
easily identifiable in the InclusiveKinematics collections

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>