HcalEndcapPInsert*Clusters: update sampFrac (#1499)
Changed sampling fraction for the insert from 0.98% to 2.59% in order to
match the change from mixed W + Fe absorbers to all Fe absorbers (see
pull request https://github.com/eic/epic/pull/744 in the epic
repository)

### Briefly, what does this PR introduce?
Changes the sampling fraction for reconstructing hits in the calorimeter
insert to match a change in the materials used in the detector

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [X] Other: single parameter tweak

### Please check if this PR fulfills the following:
- [X] Tests for the changes have been added
- [X] Documentation has been added / updated
- [X] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
yes. Users must use the updated detector geometry:
https://github.com/eic/epic/pull/744
### Does this PR change default behavior?
yes.