feat: particleSvc to distribute mass by PDG (#1487)
### Briefly, what does this PR introduce?
Instead of storing the mass of the proton etc in various classes, this
PR adds an algorithms::ParticleSvc that distributes the correct masses
and charge etc by PDG number.

Note: I was playing around with using the algorithms::ParticleSvc as a
thin interface-only, and let a JService read a data file and provide the
actual service. But I gave up since it seemed pointless.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>