Setup mc truth UndoAfterBurner (#1492)
### Briefly, what does this PR introduce?

This PR introduces the postburner to EICrecon, which will remove the
effects of the afterburner from the MCParticles branch, and store the
resulting output (the actual MC Truth information from the generator)
into a new branch, "MCParticlesHeadOnFrameNoBeamFX".

The postburner does not erase any information - the user will simply
have access to both the afterburned MC information via MCParticles (as
before), and the new branch, which handles the correct transformations
to remove the effects.

The code is also staged to be able to handle removing *only* the
crossing angle from reconstructed branches, and this additional
functionality will come with a future PR.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ x] Tests for the changes have been added
- [ x] Documentation has been added / updated
- [ x] Changes have been communicated to collaborators


### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?

[pt_DVCS_protons_MCParticles_POSTBURN_ePIC_eicrecon_6_4_2024_run_0.pdf](https://github.com/user-attachments/files/15570067/pt_DVCS_protons_MCParticles_POSTBURN_ePIC_eicrecon_6_4_2024_run_0.pdf)
[Analysis of ePIC Output With Afterburned
Events.pdf](https://github.com/user-attachments/files/15570110/Analysis.of.ePIC.Output.With.Afterburned.Events.pdf)

It only adds functionality and a new output branch, nothing else is
affected.

---------

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Simon Gardner <simon.gardner@glasgow.ac.uk>
Co-authored-by: Wouter Deconinck <wdconinc@gmail.com>
Co-authored-by: Sebouh Paul <sebouh.paul@gmail.com>