ImagingTopoCluster: fix Init printouts by adding missing break statements (#1533)
Fixed an issue where "break" is not included in a case/break statement.

### Briefly, what does this PR introduce?
When initializing the ImagingTopoCluster algorithm, there is a switch
statement wherein different things are printed to inform the user what
the parameters are (depending on the layer mode). This lacked breaks for
the cases, causing the other print statements to be executed. This patch
fixes it.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
no.  no
### Does this PR change default behavior?
yes. only prints out the information about the cuts for defining
neighbors that is relevent based on the layer mode.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>