Added missing seed finder parameter (#1544)
### Briefly, what does this PR introduce?
The _rMin_ parameter defined in the seeder configuration file was not
being passed to the seed finder. This PR adds this. The parameter is
currently set to the default value defined here:


https://github.com/acts-project/acts/blob/main/Core/include/Acts/Seeding/SeedFinderOrthogonalConfig.hpp#L39

https://github.com/acts-project/acts/blob/main/Core/include/Acts/Seeding/SeedFinderConfig.hpp#L43

So, there should be no change to the output. But it will be needed in
the future if we ever want to search for seeds in an outer part of the
tracking detector volume.

In addition, the _bFieldMin_ parameter defined in the seeder
configuration file but not passed to the seed finder has been removed.
This parameter does not seem to be a Acts seed finder parameter in any
event. It seems to only be used if the _EstimateTrackParamsFromSeed_
class is called to fit the seed parameters (which we do not use):


https://github.com/acts-project/acts/blob/main/Core/include/Acts/Seeding/EstimateTrackParamsFromSeed.hpp#L148-L149


### What kind of change does this PR introduce?
- [x] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.